+++ network.h
@@ -21,11 +21,11 @@ void network_init_backprop(network const *const net);
 void network_free(network *const net);
 
 double **network_pass_forward(
-    network const *const net, size_t t, uint8_t *const *const images);
+    network const *const net, size_t t, double *const *const images);
 
 void network_train(
     network const *const net, size_t epochs, size_t t,
-    uint8_t *const *const images, uint8_t *const labels);
+    double *const *const images, uint8_t *const labels);
 
 void network_save_results(
     char const *const result_fname, size_t t, double *const *const result);
